Added README instructions, moved code from __init__
